+++ CesiumRasterOverlays/include/CesiumRasterOverlays/IonRasterOverlay.h
@@ -10,10 +10,24 @@
 
 namespace CesiumRasterOverlays {
 
+/**
+ * @brief A class specifying how an endpoint should be accessed.
+ */
 class EndpointResource {
 public:
+  /**
+   * @brief If this method returns true, the access token should be included in
+   * the `Authorization` header of the initial request.
+   */
   virtual bool needsAuthHeaderOnInitialRequest() const = 0;
 
+  /**
+   * @brief Returns the endpoint URL for the given asset.
+   *
+   * @param assetID The ID of the asset to obtain.
+   * @param accessToken The access token to use to obtain the asset.
+   * @param assetEndpointUrl The base URL of the asset endpoint.
+   */
   virtual std::string getUrl(
       int64_t assetID,
       const std::string& accessToken,
@@ -59,6 +73,22 @@ public:
       const override;
 
 protected:
+  /**
+   * @brief Creates a new instance.
+   *
+   * The tiles that are provided by this instance will contain
+   * imagery data that was obtained from the Cesium ion asset
+   * with the given ID, accessed with the given access token.
+   *
+   * @param name The user-given name of this overlay layer.
+   * @param ionAssetID The asset ID.
+   * @param ionAccessToken The access token.
+   * @param endpointResource The \ref EndpointResource to use to access the ion
+   * API or an ion-compatible API.
+   * @param overlayOptions The {@link RasterOverlayOptions} for this instance.
+   * @param ionAssetEndpointUrl The URL of the ion endpoint to make our requests
+   * to.
+   */
   IonRasterOverlay(
       const std::string& name,
       int64_t ionAssetID,
