+++ doc/diagrams/dependencies/CesiumClientCommon.mmd
@@ -1,0 +1,16 @@
+---
+comment: This file was generated by dep-graph-gen. DO NOT EDIT THIS FILE!
+
+title: CesiumClientCommon Dependency Graph
+---
+graph TD
+  classDef dependencyNode fill:#fff,stroke:#ccc,color:#666,font-weight:bold,font-size:28px
+  classDef libraryNode fill:#9f9,font-weight:bold,font-size:28px
+  CesiumClientCommon[CesiumClientCommon] --> CesiumAsync[CesiumAsync]
+  CesiumClientCommon[CesiumClientCommon] --> CesiumUtility[CesiumUtility]
+  CesiumClientCommon[CesiumClientCommon] --> OpenSSL_Crypto{{OpenSSL::Crypto}}
+  CesiumClientCommon[CesiumClientCommon] --> httplib_httplib{{httplib::httplib}}
+  CesiumClientCommon[CesiumClientCommon] --> modp_b64_modp_b64{{modp_b64::modp_b64}}
+  CesiumClientCommon[CesiumClientCommon] --> picosha2_picosha2{{picosha2::picosha2}}
+  class OpenSSL_Crypto,httplib_httplib,modp_b64_modp_b64,picosha2_picosha2 dependencyNode
+  class CesiumAsync,CesiumUtility,CesiumClientCommon libraryNode
