+++ doc/diagrams/dependencies/CesiumITwinClient.mmd
@@ -1,0 +1,15 @@
+---
+comment: This file was generated by dep-graph-gen. DO NOT EDIT THIS FILE!
+
+title: CesiumITwinClient Dependency Graph
+---
+graph TD
+  classDef dependencyNode fill:#fff,stroke:#ccc,color:#666,font-weight:bold,font-size:28px
+  classDef libraryNode fill:#9f9,font-weight:bold,font-size:28px
+  CesiumITwinClient[CesiumITwinClient] --> CesiumAsync[CesiumAsync]
+  CesiumITwinClient[CesiumITwinClient] --> CesiumClientCommon[CesiumClientCommon]
+  CesiumITwinClient[CesiumITwinClient] --> CesiumGeospatial[CesiumGeospatial]
+  CesiumITwinClient[CesiumITwinClient] --> CesiumUtility[CesiumUtility]
+  CesiumITwinClient[CesiumITwinClient] --> modp_b64_modp_b64{{modp_b64::modp_b64}}
+  class modp_b64_modp_b64 dependencyNode
+  class CesiumAsync,CesiumClientCommon,CesiumGeospatial,CesiumUtility,CesiumITwinClient libraryNode
